veinbot/veinbot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
psutil>=5.9.0
Pillow>=10.0.0
openai>=1.0.0
# Optional — only needed if you use music (requires Lavalink running separately):
# wavelink>=3.0.0
